requirements.txt
streamlit
Pillow
numpy
streamlit-drawable-canvas
opencv-python-headless==4.5.5.62
tensorflow
keras